backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
langchain-huggingface
chromadb
python-dotenv
tiktoken
sentence-transformers
youtube-transcript-api
pydantic
torch
numpy
requests
httpx